requirements.txt
paddlepaddle==3.2.2
paddleocr==3.3.2
# PaddleX OCR 依赖组（用于印章文本识别产线）
# paddlex[ocr]>=3.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
PyYAML>=6.0
Flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
loguru>=0.7.0